+++ data/onboarding/build.gradle.kts
@@ -1,0 +1,9 @@
+import com.hilingual.buildlogic.setNamespace
+
+plugins {
+    alias(libs.plugins.hilingual.android.data)
+}
+
+android {
+    setNamespace("data.onboarding")
+}
