+++ data/onboarding/src/main/java/com/hilingual/data/onboarding/repository/OnboardingRepository.kt
@@ -1,0 +1,6 @@
+package com.hilingual.data.onboarding.repository
+
+interface OnboardingRepository {
+    suspend fun getIsHomeOnboardingCompleted(): Result<Boolean>
+    suspend fun updateIsHomeOnboardingCompleted(isCompleted: Boolean): Result<Unit>
+}
