+++ data/user/src/main/java/com/hilingual/data/user/repository/UserRepository.kt
@@ -82,7 +82,4 @@ interface UserRepository {
     suspend fun deleteFollow(
         targetUserId: Long
     ): Result<Unit>
-
-    suspend fun getIsHomeOnboardingCompleted(): Result<Boolean>
-    suspend fun updateIsHomeOnboardingCompleted(isCompleted: Boolean): Result<Unit>
 }
